requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
datetime-truncate==1.1.1
dj-database-url==2.1.0
Django==5.0.4
django-admin-charts==1.3.1
django-admin-sortable==2.3
django-admin-sortable2==2.1.10
django-admin-tools==0.9.3
django-admin-tools-stats==0.9.0
django-bower==5.2.0
django-chartit==0.2.9
django-ckeditor==6.7.1
django-grappelli==4.0.1
django-js-asset==2.2.0
django-jsonfield==1.4.1
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.10.1
django-qsstats-magic==1.1.0
django-unfold==0.22.0
djcacheutils==3.0.0
geoip2==2.9.0
git-filter-repo==2.38.0
gunicorn==22.0.0
h11==0.14.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
maxminddb==2.6.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-memcached==1.62
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
whitenoise==6.6.0
